+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.12.0-beta.2
+# Added
+- Support ABI reference types and other improvements (#273)
+# Changed
+- Fix wait for confirmation function (#267)
 # 1.12.0-beta.1
 # Added
 - EncodeAddress